requirements.txt
streamlit
python-docx
python-dotenv
openai
sentence-transformers
faiss-cpu
PyPDF2
tqdm
numpy
